+++ src/streaming/downsampler.hh
@@ -51,11 +51,8 @@ class Downsampler
     std::unordered_map<int, ByteVector> downsampled_frames_;
     std::unordered_map<int, ByteVector> partial_scaled_frames_;
 
-    bool is_3d_downsample_() const;
     size_t n_levels_() const;
 
     void make_writer_configurations_(std::shared_ptr<ArrayConfig> config);
-    void downsample_3d_(ConstByteSpan frame_data);
-    void downsample_2d_(ConstByteSpan frame_data);
 };
 } // namespace zarr